# speed up polynomial computation